Updated index.js and created color-operations.js
