+++ color-operations.js
@@ -1,0 +1,44 @@
+const MajorColors = [
+    "white",
+    "red",
+    "black",
+    "yellow",
+    "violet"
+];
+
+const MinorColors = [
+    "blue",
+    "orange",
+    "green",
+    "brown",
+    "slate"
+];
+
+function GetColorFromPairNumber(pairNumber) {
+    let colorPair = {};
+    const zeroBasedPairNumber = pairNumber - 1;
+    const majorColorIndex =
+        Math.floor(zeroBasedPairNumber / MinorColors.length);
+    colorPair.major = MajorColors[majorColorIndex];
+    colorPair.minor = MinorColors[zeroBasedPairNumber % MinorColors.length];
+    return colorPair
+}
+
+function GetPairNumberFromColors(majorColor, minorColor) {
+    let majorIndex = 0;
+    let minorIndex = 0;
+    for(majorIndex = 0; majorIndex < MajorColors.length; majorIndex++) {
+        if(MajorColors[majorIndex] == majorColor) {
+            break;
+        }
+    }
+    for(minorIndex = 0; minorIndex < MinorColors.length; minorIndex++) {
+        if(MinorColors[minorIndex] == minorColor) {
+            break;
+        }
+    }
+    return majorIndex * MinorColors.length + minorIndex + 1;
+}
+
+module.exports = GetColorFromPairNumber;
+module.exports = GetPairNumberFromColors;
